requirements.txt
tensorflow-macos
tensorflow-metal
tflite-model-maker
librosa
soundfile
seaborn
matplotlib